rag_app/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
chromadb>=0.4.18
pypdf>=3.17.0
python-docx>=1.1.0
sentence-transformers>=2.2.2
pydantic>=2.5.0